// interfaces/orderfulfillment.interface.ts
export interface IOrderFulfillment {
    bookCode?: string;
    namePlate?: string;
    marketingGroup?: string;
    optionGroup?: string;
    description?: string;
    modelYear?: string;
    geographicMarket?: string;
    vehicleLine?: string;
    effectiveDate?: string;
    modelYearSuffix?: string;
}